Fixed SignalR connection
Swapped out the @aspnet/signalr npm package for the most current
@microsoft/signalr package. I do NOT think this had anything to do with
fixing the connection. However, I was definitely using an outdated
package.

The fix had to do with the url I was trying to connect with, which makes
a lot of sense in retrospect. When the signalr hub is created during
the API startup, it's assigned its own path. The client was pointing at
the messages api path rather than the signalr hub path when connecting
to signalr. Changing this immediately resolved all connection errors.